requirements.txt
dislash.py==1.4.9
qrcode==7.1
pytz==2018.3
web3==5.21.0
discord.py==1.7.3
numpy==1.19.5
eth_account==0.5.4
pandas==1.1.5
requests==2.18.4
urllib3==1.22
beautifulsoup4==4.10.0
discord==1.7.3
Pillow==8.3.2
twocaptcha==0.0.1
